Ditch the duplication MinecraftJarProvider has with MinecraftProvider
They both declare the merged jar location and it is mere luck that it's the same place
